requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
gunicorn==21.2.0

# LangChain and LLM
langchain>=0.1.0,<0.3.0
langchain-google-genai>=0.0.6

# Database
sqlalchemy==2.0.23

# Security
bcrypt>=4.0.0,<5.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator>=2.0.0

# Environment
python-dotenv==1.0.0

# Image Generation
google-genai<=0.4.0
pillow>=10.2.0,<11.0.0

# File Processing
pypdf2==3.0.1
python-docx==1.1.0

# Utilities
requests==2.31.0

# Rate Limiting
slowapi==0.1.9

# Logging
python-json-logger==2.0.7

